커스텀 포매터ㅣdamage는 이제 int로 저장됩니다.

+++ RGM/RGM/EventArgs/PlayerEvents.cs
@@ -535,7 +535,7 @@ namespace RGM.EventArgs
                 else
                 {
                     if (ev.IsAllowed)
-                        PlayersReport[ev.Attacker.UserId].Damage += ev.DamageHandler.Damage;
+                        PlayersReport[ev.Attacker.UserId].Damage += (int)ev.DamageHandler.Damage;
                 }
             }
         }
